requirements.txt
torch==2.5.1
torchaudio==2.5.1
torchvision==2.5.1
wxpython
pandas
pypubsub
pyyaml
scikit-learn
pyteomics
statsmodels
ropwr
pulp
lxml
tensorboard
lightning
scipy
matplotlib
gputil
ecos
numpy==1.26.4
pyinstaller==6.2.0